Make it possible to use double precision shifts
Until now, pixel shiftings were expressed as integers. However, because
we're doing interpolation and that the precision of measurements is sub-pixel,
we can definitely use double precision shifts. This opens the ability to
generate much smoother animations of the limb, for example.

+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/event/ImageLine.java
@@ -15,5 +15,5 @@
  */
 package me.champeau.a4j.jsolex.processing.event;
 
-public record ImageLine(int pixelShift, int line, int totalLines, double[] data, boolean display) {
+public record ImageLine(double pixelShift, int line, int totalLines, double[] data, boolean display) {
 }
